+++ R/calculateModuleData.R
@@ -111,7 +111,7 @@ calculateModuleData <- function(binData, binFuns=list('ClickDetector'=list(stand
                                stringsAsFactors = FALSE)
                 })) %>%
                 mutate(detectorName = paste(detName, .data$detectorName, sep='_'))
-            allClicks <- left_join(allClicks, allNames, by='UID')
+            allClicks <- left_join(allClicks, allNames, join_by('UID'=='UID'), multiple='all')
             allClicks$callType <- 'click'
             allClicks
         },
@@ -269,7 +269,7 @@ doCalcs <- function(data, funs, detSettings=NULL, module, retry=TRUE) {
                    for(j in seq_along(times)) {
                        whichThis <- allTimes == times[j]
                        maxEn <- max(thisEnergy[whichThis])
-                       whichMax <- whichThis & (thisEnergy == maxEn)
+                       whichMax <- which(whichThis & (thisEnergy == maxEn))[1]
                        freq[j] <- thisFreq[whichMax]
                    }
                    data[[i]]$freq <- freq