+++ R/getDetectorData.R
@@ -57,6 +57,9 @@ getDetectorData <- function(x, measures=TRUE) {
     }
     # base case one acev
     dets <- detectors(x)
+    if(length(dets) == 0) {
+        return(NULL)
+    }
     callTypes <- getCallType(dets)
     meas <- ancillary(x)$measures
     for(d in seq_along(dets)) {
@@ -144,9 +147,10 @@ getMeasures <- function(x) {
         result <- lapply(x[sapply(x, is.AcousticEvent)], function(e) {
             getMeasures(e)
         })
-        names(result) <- NULL
-        result <- unlist(result, recursive=FALSE)
-        return(squishList(result))
+        return(bind_rows(result))
+        # names(result) <- NULL
+        # result <- unlist(result, recursive=FALSE)
+        # return(squishList(result))
     }
     # base case one acev
     c(id=id(x), ancillary(x)$measures)